requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==23.1.0
av==10.0.0
beautifulsoup4==4.12.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.4.1
entrypoints==0.4
faiss-cpu==1.9.0
Faker==33.0.0
favicon==0.7.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.5.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
htbuilder==0.7.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.4
importlib-metadata==6.8.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.7.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.13.1
kiwisolver==1.4.5
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langsmith==0.1.143
lazy_loader==0.3
libclang==16.0.6
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.23.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.54.4
opencv-python==4.8.0.74
opt-einsum==3.3.0
orjson==3.10.11
packaging==24.2
pandas==2.1.2
Pillow==10.0.0
plotly==5.24.1
prometheus_client==0.21.0
propcache==0.2.0
protobuf==4.23.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.16.1
pylibsrtp==0.8.0
pymdown-extensions==10.12
pymodbus==3.4.1
pymongo==4.4.1
pyodbc==4.0.39
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf==5.1.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.40.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit-webrtc==0.47.1
TBB==0.2
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.18
tiktoken==0.8.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.16
utils==1.0.1
validators==0.22.0
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.17.1
zipp==3.17.0
